:recycle: #54 Router内のロジックをControllerに移動
